* commit '62db2a9eb111366a2331fa722b5568e7d06e0b61':
  правка доктеста
  правка в доке
  нембольшие правки
  добавил доки немного
  Поправил замечания
  from ignite.engine import Events
  правка теста в utils
  Убрал лишнюю функцию
  тесты и комменты
  admmslim

+++ sponge_bob_magic/utils.py
@@ -187,7 +187,7 @@ def to_csr(
     >>> m = to_csr(data_frame)
     >>> m.toarray()
     array([[1, 0, 0],
-           [0, 0, 2]], dtype=int64)
+           [0, 0, 2]])
 
     :param log: spark DataFrame с колонками ``user_id``, ``item_id`` и ``relevance``
     :param user_count: количество строк в результирующей матрице (если пусто, то вычисляется по логу)
